requirements.txt
# Core scientific computing
numpy>=1.23.5,<1.25.0
pandas>=1.5.2,<2.0.0
scipy>=1.9.3,<1.12.0

# Machine learning
scikit-learn>=1.3.0,<1.4.0
optuna>=4.3.0,<4.5.0
shap>=0.41.0,<0.43.0
xgboost>=1.5.0,<2.0.0

# Data handling and I/O
netcdf4>=1.5.7,<1.7.0
h5py>=3.7.0,<3.10.0
joblib>=1.1.1,<1.4.0

# Visualization
matplotlib>=3.5.0,<3.8.0
seaborn>=0.12.2,<0.13.0
cartopy>=0.21.1,<0.22.0

# Image processing (for tools)
pillow>=9.3.0,<10.0.0

# Geospatial
pyproj>=3.3.1,<3.7.0

# Progress bars
tqdm>=4.64.1,<5.0.0

# Performance
numba>=0.56.4,<0.59.0
